% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_.R
\name{add_}
\alias{add_}
\alias{add_occurrences}
\alias{add_events}
\alias{add_multimedia}
\alias{add_metadata}
\alias{add_data}
\alias{add_schema}
\title{Add information to a \code{darwincore} object}
\usage{
add_occurrences(.dwc, df = NULL)

add_events(.dwc, df = NULL)

add_multimedia(.dwc, df = NULL)

add_metadata(.dwc, xml = NULL)

add_data(.dwc, df = NULL, slot = NULL)

add_schema(.dwc)
}
\arguments{
\item{.dwc}{An object of class \code{darwincore}, constructed with \code{darwincore()}.}

\item{df}{Object of class \code{tibble} or \code{data.frame}.}

\item{xml}{Object of class \code{xml_document}, such as imported from
\code{.xml} with \code{xml2::read_xml()} or from markdown (\code{.rmd}) using
\code{galaxias::read_md()}.}

\item{slot}{Name of the slot where this data should go.}
}
\value{
An object of class \code{dwca}, updated with the provided information.
}
\description{
These are all pipeable functions for adding relevant information to a
\code{darwincore} object. The underlying workhorse function is \code{add_data()}, but
it is safer to call the more specific functions \code{add_occurrences()},
\code{add_events()} and \code{add_media()}. \code{add_metadata()} is also mandatory and
requires either an \code{xml} object/file or a markdown file (\code{rmd} or \code{md}
preferred). \code{add_schema()} is called internally and there is no
requirement for the user to call it, but it is shipped here as it may be
useful for debugging purposes.
}
