% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwincore-package.R
\docType{package}
\name{darwincore-package}
\alias{darwincore-package}
\title{Object class for storing Darwin Core Archives in R}
\description{
\code{{darwincore}} is an functionally similar to \code{{finch}}. It provides tools to
import Darwin Core Archives in an R-native format; offers summary, print, and
conversion functions for those objects; and gives tools to read and
incorporate external \code{.csv} and markdown (\code{.md}) objects into those archives.
}
\section{Functions}{


\strong{Object manipulation}
\itemize{
\item \code{\link[=darwincore]{darwincore()}} create a \code{darwincore} object
\item \code{\link[=add_data]{add_data()}}
\item \code{\link[=add_metadata]{add_metadata()}}
\item \code{\link[=print.darwincore]{print.darwincore()}}, \code{\link[=as_tibble.darwincore]{as_tibble.darwincore()}}, \code{\link[=summary.darwincore]{summary.darwincore()}}
}

\strong{File manipulation}
\itemize{
\item \code{\link[=read_dwca]{read_dwca()}} read a Darwin Core Archive as a \code{darwincore} object
\item \code{\link[=write_dwca]{write_dwca()}} write a \code{darwincore} object to a Darwin Core Archive
\item \code{\link[=read_md]{read_md()}} read a markdown file as xml
\item \code{\link[=write_md]{write_md()}} convert an xml document to a markdown file
}
}

\references{
If you have any questions, comments or suggestions, please email
\href{mailto:support@ala.org.au}{support@ala.org.au}.
}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Shandiya Balasubramanium \email{shandiya.balasubramanium@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
}

}
\keyword{internal}
