% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darwincore-class.R
\name{darwincore}
\alias{darwincore}
\alias{print.darwincore}
\title{Create an object of class \code{darwincore}}
\usage{
darwincore()

\method{print}{darwincore}(x, ...)
}
\value{
Objects of class \code{darwincore} are listlike, and contain two slots:
\code{data} and \code{metadata}. The former contains and supplied data. The latter
contains \code{statement} and \code{schema}.
}
\description{
Create an object of class \code{darwincore}
}
