% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dwca.R
\name{read_dwca}
\alias{read_dwca}
\title{Read a Darwin Core Archive as a \code{darwincore} object}
\usage{
read_dwca(file)
}
\description{
Read a Darwin Core Archive as a \code{darwincore} object
}
\seealso{
\code{\link[=write_dwca]{write_dwca()}}
}
