% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dwca.R
\name{write_dwca}
\alias{write_dwca}
\title{Write \code{darwincore} objects to a Darwin Core Archive}
\usage{
write_dwca(.dwca, file = NULL)
}
\arguments{
\item{.dwc}{A \code{darwincore} object}

\item{path}{(Optional) Name of the zip file. Defaults to \code{NULL}, indicating
that the file in question should be saved to a temporary directory.}
}
\value{
\code{write_dwca()} is called for the side-effect of building a 'Darwin
Core Archive' (i.e. a zip file).
}
\description{
\code{write_dwca()} converts a \code{darwincore} archive to a zip file.
}
\seealso{
\code{\link[=read_dwca]{read_dwca()}}
}
